+++ pytorch_forecasting/models/base_model.py
@@ -6,6 +6,7 @@ import copy
 from copy import deepcopy
 import inspect
 from typing import Any, Callable, Dict, Iterable, List, Tuple, Union
+import warnings
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -36,9 +37,11 @@ from pytorch_forecasting.metrics import (
     QuantileLoss,
     convert_torchmetric_to_pytorch_forecasting_metric,
 )
+from pytorch_forecasting.models.nn.embeddings import MultiEmbedding
 from pytorch_forecasting.optim import Ranger
 from pytorch_forecasting.utils import (
     OutputMixIn,
+    TupleOutputMixIn,
     apply_to_list,
     create_mask,
     get_embedding_size,
@@ -73,7 +76,24 @@ def _torch_cat_na(x: List[torch.Tensor]) -> torch.Tensor:
                 )
                 for xi in x
             ]
-    return torch.cat(x, dim=0)
+
+    # check if remaining dimensions are all equal
+    if x[0].ndim > 2:
+        remaining_dimensions_equal = all([all([xi.size(i) == x[0].size(i) for xi in x]) for i in range(2, x[0].ndim)])
+    else:
+        remaining_dimensions_equal = True
+
+    # deaggregate
+    if remaining_dimensions_equal:
+        return torch.cat(x, dim=0)
+    else:
+        # make list instead but warn
+        warnings.warn(
+            f"Not all dimensions are equal for tensors shapes. Example tensor {x[0].shape}. "
+            "Returning list instead of torch.Tensor.",
+            UserWarning,
+        )
+        return [xii for xi in x for xii in xi]
 
 
 def _concatenate_output(
@@ -131,7 +151,7 @@ STAGE_STATES = {
 }
 
 
-class BaseModel(LightningModule):
+class BaseModel(LightningModule, TupleOutputMixIn):
     """
     BaseModel from which new timeseries models should inherit from.
     The ``hparams`` of the created object will default to the parameters indicated in :py:meth:`~__init__`.
@@ -192,6 +212,7 @@ class BaseModel(LightningModule):
         loss: Metric = SMAPE(),
         logging_metrics: nn.ModuleList = nn.ModuleList([]),
         reduce_on_plateau_patience: int = 1000,
+        reduce_on_plateau_reduction: float = 2.0,
         reduce_on_plateau_min_lr: float = 1e-5,
         weight_decay: float = 0.0,
         optimizer_params: Dict[str, Any] = None,
@@ -215,6 +236,7 @@ class BaseModel(LightningModule):
                 Defaults to [].
             reduce_on_plateau_patience (int): patience after which learning rate is reduced by a factor of 10. Defaults
                 to 1000
+            reduce_on_plateau_reduction (float): reduction in learning rate when encountering plateau. Defaults to 2.0.
             reduce_on_plateau_min_lr (float): minimum learning rate for reduce on plateua learning rate scheduler.
                 Defaults to 1e-5
             weight_decay (float): weight decay. Defaults to 0.0.
@@ -430,6 +452,12 @@ class BaseModel(LightningModule):
             Dict[str, Any]: log dictionary to be returned by training and validation steps
         """
         # log
+        if isinstance(self.loss, DistributionLoss):
+            prediction_kwargs.setdefault("n_samples", 20)
+            prediction_kwargs.setdefault("use_metric", True)
+            quantiles_kwargs.setdefault("n_samples", 20)
+            quantiles_kwargs.setdefault("use_metric", True)
+
         self.log_metrics(x, y, out, prediction_kwargs=prediction_kwargs)
         if self.log_interval > 0:
             self.log_prediction(
@@ -513,7 +541,7 @@ class BaseModel(LightningModule):
             # multiply monotinicity loss by large number to ensure relevance and take to the power of 2
             # for smoothness of loss function
             monotinicity_loss = 10 * torch.pow(monotinicity_loss, 2)
-            if isinstance(self.loss, MASE):
+            if isinstance(self.loss, (MASE, MultiLoss)):
                 loss = self.loss(
                     prediction, y, encoder_target=x["encoder_target"], encoder_lengths=x["encoder_lengths"]
                 )
@@ -526,10 +554,9 @@ class BaseModel(LightningModule):
 
             # calculate loss
             prediction = out["prediction"]
-            if isinstance(self.loss, MASE):
-                loss = self.loss(
-                    prediction, y, encoder_target=x["encoder_target"], encoder_lengths=x["encoder_lengths"]
-                )
+            if isinstance(self.loss, (MASE, MultiLoss)):
+                mase_kwargs = dict(encoder_target=x["encoder_target"], encoder_lengths=x["encoder_lengths"])
+                loss = self.loss(prediction, y, **mase_kwargs)
             else:
                 loss = self.loss(prediction, y)
 
@@ -594,27 +621,6 @@ class BaseModel(LightningModule):
                     on_epoch=True,
                     batch_size=len(x["decoder_target"]),
                 )
-
-    def to_network_output(self, **results):
-        """
-        Convert output into a named (and immuatable) tuple.
-
-        This allows tracing the modules as graphs and prevents modifying the output.
-
-        Returns:
-            named tuple
-        """
-        if hasattr(self, "_output_class"):
-            Output = self._output_class
-        else:
-            OutputTuple = namedtuple("output", results)
-
-            class Output(OutputMixIn, OutputTuple):
-                pass
-
-            self._output_class = Output
-
-        return self._output_class(**results)
 
     def forward(
         self, x: Dict[str, Union[torch.Tensor, List[torch.Tensor]]]
@@ -730,8 +736,8 @@ class BaseModel(LightningModule):
         add_loss_to_title: Union[Metric, torch.Tensor, bool] = False,
         show_future_observed: bool = True,
         ax=None,
-        quantiles_kwargs: Dict[str, Any] = None,
-        prediction_kwargs: Dict[str, Any] = None,
+        quantiles_kwargs: Dict[str, Any] = {},
+        prediction_kwargs: Dict[str, Any] = {},
     ) -> plt.Figure:
         """
         Plot prediction of prediction vs actuals
@@ -756,11 +762,9 @@ class BaseModel(LightningModule):
         decoder_targets = to_list(x["decoder_target"])
 
         # get predictions
-        if prediction_kwargs is None:
-            prediction_kwargs = {}
-
-        if quantiles_kwargs is None:
-            quantiles_kwargs = {}
+        if isinstance(self.loss, DistributionLoss):
+            prediction_kwargs.setdefault("use_metric", False)
+            quantiles_kwargs.setdefault("use_metric", False)
 
         y_raws = to_list(out["prediction"])  # raw predictions - used for calculating loss
         y_hats = to_list(self.to_prediction(out, **prediction_kwargs))
@@ -866,7 +870,7 @@ class BaseModel(LightningModule):
         for name, p in named_parameters:
             if p.grad is not None and p.requires_grad and "bias" not in name:
                 layers.append(name)
-                ave_grads.append(p.grad.abs().mean())
+                ave_grads.append(p.grad.abs().cpu().mean())
                 self.logger.experiment.add_histogram(tag=name, values=p.grad, global_step=self.global_step)
         fig, ax = plt.subplots()
         ax.plot(ave_grads)
@@ -940,7 +944,7 @@ class BaseModel(LightningModule):
             raise ValueError(f"Optimizer of self.hparams.optimizer={self.hparams.optimizer} unknown")
 
         # set scheduler
-        if isinstance(lr, (list, tuple)):  # change for each epoch
+        if isinstance(lrs, (list, tuple)):  # change for each epoch
             # normalize lrs
             lrs = np.array(lrs) / lrs[0]
             scheduler_config = {
@@ -956,7 +960,7 @@ class BaseModel(LightningModule):
                 "scheduler": ReduceLROnPlateau(
                     optimizer,
                     mode="min",
-                    factor=0.2,
+                    factor=1.0 / self.hparams.reduce_on_plateau_reduction,
                     patience=self.hparams.reduce_on_plateau_patience,
                     cooldown=self.hparams.reduce_on_plateau_patience,
                     min_lr=self.hparams.reduce_on_plateau_min_lr,
@@ -1348,6 +1352,25 @@ class BaseModelWithCovariates(BaseModel):
     """
 
     @property
+    def target_positions(self) -> torch.LongTensor:
+        """
+        Positions of target variable(s) in covariates.
+
+        Returns:
+            torch.LongTensor: tensor of positions.
+        """
+        # todo: expand for categorical targets
+        if "target" in self.hparams:
+            target = self.hparams.target
+        else:
+            target = self.dataset_parameters["target"]
+        return torch.tensor(
+            [self.hparams.x_reals.index(name) for name in to_list(target)],
+            device=self.device,
+            dtype=torch.long,
+        )
+
+    @property
     def reals(self) -> List[str]:
         """List of all continuous variables in model"""
         return list(
@@ -1453,6 +1476,47 @@ class BaseModelWithCovariates(BaseModel):
         )
         new_kwargs.update(kwargs)
         return super().from_dataset(dataset, **new_kwargs)
+
+    def extract_features(
+        self,
+        x,
+        embeddings: MultiEmbedding = None,
+        period: str = "all",
+    ) -> torch.Tensor:
+        """
+        Extract features
+
+        Args:
+            x (Dict[str, torch.Tensor]): input from the dataloader
+            embeddings (MultiEmbedding): embeddings for categorical variables
+            period (str, optional): One of "encoder", "decoder" or "all". Defaults to "all".
+
+        Returns:
+            torch.Tensor: tensor with selected variables
+        """
+        # select period
+        if period == "encoder":
+            x_cat = x["encoder_cat"]
+            x_cont = x["encoder_cont"]
+        elif period == "decoder":
+            x_cat = x["decoder_cat"]
+            x_cont = x["decoder_cont"]
+        elif period == "all":
+            x_cat = torch.cat([x["encoder_cat"], x["decoder_cat"]], dim=1)  # concatenate in time dimension
+            x_cont = torch.cat([x["encoder_cont"], x["decoder_cont"]], dim=1)  # concatenate in time dimension
+        else:
+            raise ValueError(f"Unknown type: {type}")
+
+        # create dictionary of encoded vectors
+        input_vectors = embeddings(x_cat)
+        input_vectors.update(
+            {
+                name: x_cont[..., idx].unsqueeze(-1)
+                for idx, name in enumerate(self.hparams.x_reals)
+                if name in self.reals
+            }
+        )
+        return input_vectors
 
     def calculate_prediction_actual_by_variable(
         self,
@@ -1740,6 +1804,7 @@ class AutoRegressiveBaseModel(BaseModel):
         self,
         normalized_prediction_parameters: torch.Tensor,
         target_scale: Union[List[torch.Tensor], torch.Tensor],
+        n_samples: int = 1,
         **kwargs,
     ) -> Tuple[Union[List[torch.Tensor], torch.Tensor], torch.Tensor]:
         """
@@ -1750,6 +1815,7 @@ class AutoRegressiveBaseModel(BaseModel):
         Args:
             normalized_prediction_parameters (torch.Tensor): network prediction output
             target_scale (Union[List[torch.Tensor], torch.Tensor]): target scale to rescale network output
+            n_samples (int, optional): Number of samples to draw independently. Defaults to 1.
             **kwargs: extra arguments for dictionary passed to :py:meth:`~transform_output` method.
 
         Returns:
@@ -1769,7 +1835,15 @@ class AutoRegressiveBaseModel(BaseModel):
             isinstance(self.loss, MultiLoss) and isinstance(self.loss[0], DistributionLoss)
         ):
             # todo: handle mixed losses
-            prediction = self.loss.sample(prediction_parameters, 1)
+            if n_samples > 1:
+                prediction_parameters = apply_to_list(
+                    prediction_parameters, lambda x: x.reshape(int(x.size(0) / n_samples), n_samples, -1)
+                )
+                prediction = self.loss.sample(prediction_parameters, 1)
+                prediction = apply_to_list(prediction, lambda x: x.reshape(x.size(0) * n_samples, 1, -1))
+            else:
+                prediction = self.loss.sample(normalized_prediction_parameters, 1)
+
         else:
             prediction = prediction_parameters
         # normalize prediction prediction
@@ -1792,6 +1866,7 @@ class AutoRegressiveBaseModel(BaseModel):
         first_hidden_state: Any,
         target_scale: Union[List[torch.Tensor], torch.Tensor],
         n_decoder_steps: int,
+        n_samples: int = 1,
         **kwargs,
     ) -> Union[List[torch.Tensor], torch.Tensor]:
         """
@@ -1817,6 +1892,8 @@ class AutoRegressiveBaseModel(BaseModel):
             first_hidden_state (Any): first hidden state used for decoding
             target_scale (Union[List[torch.Tensor], torch.Tensor]): target scale as in ``x``
             n_decoder_steps (int): number of decoding/prediction steps
+            n_samples (int): number of independent samples to draw from the distribution -
+                only relevant for multivariate models. Defaults to 1.
             **kwargs: additional arguments that are passed to the decode_one function.
 
         Returns:
@@ -1910,7 +1987,9 @@ class AutoRegressiveBaseModel(BaseModel):
             )
 
             # get prediction and its normalized version for the next step
-            prediction, current_target = self.output_to_prediction(current_target, target_scale=target_scale)
+            prediction, current_target = self.output_to_prediction(
+                current_target, target_scale=target_scale, n_samples=n_samples
+            )
             # save normalized output for lagged targets
             normalized_output.append(current_target)
             # set output to unnormalized samples, append each target as n_batch_samples x n_random_samples
@@ -1982,21 +2061,6 @@ class AutoRegressiveBaseModelWithCovariates(BaseModelWithCovariates, AutoRegress
             can also take multiple values simultaneously (e.g. holiday during octoberfest). They should be implemented
             as bag of embeddings
     """
-
-    @property
-    def target_positions(self) -> torch.LongTensor:
-        """
-        Positions of target variable(s) in covariates.
-
-        Returns:
-            torch.LongTensor: tensor of positions.
-        """
-        # todo: expand for categorical targets
-        return torch.tensor(
-            [self.hparams.x_reals.index(name) for name in to_list(self.hparams.target)],
-            device=self.device,
-            dtype=torch.long,
-        )
 
     @property
     def lagged_target_positions(self) -> Dict[int, torch.LongTensor]: